+++ ouroboros-consensus.cabal
@@ -1833,6 +1833,7 @@ library unstable-cardano-tools
     Cardano.Tools.DBAnalyser.Block.Shelley
     Cardano.Tools.DBAnalyser.CSV
     Cardano.Tools.DBAnalyser.HasAnalysis
+    Cardano.Tools.DBAnalyser.Leios
     Cardano.Tools.DBAnalyser.Run
     Cardano.Tools.DBAnalyser.Types
     Cardano.Tools.DBImmutaliser.Run
